added glass_tank & petting_area
