+++ models/__init__.py
@@ -1,1 +1,3 @@
-from .animals import ShireHorse, Rattlesnake
+from models.glass_tank.slithering_critters import  Rattlesnake, RoughGreenSnake, MilkSnake, GopherSnake, GreenSnake
+from models.petting_area.walking_critters import ShireHorse, ClydesdaleHorse, TamworthPig, HampshirePig, Llama
+from models.pond.swimming_critters import Goldfish, Mallard, RainbowTrout, Bluegrill, Pumpkinseed
