+++ models/pond/swimming_critters.py
@@ -1,0 +1,56 @@
+
+# python datetime module to help us create a timestamp
+from datetime import date
+
+
+
+class Goldfish:
+
+    def __init__(self):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = "Goldy"
+        self.species = "fish"
+        self.swimming = True
+        self.date_added =  date.today()
+        
+class Mallard:
+
+    def __init__(self):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = "Mason"
+        self.species = "fish"
+        self.swimming = True
+        self.date_added =  date.today()
+        
+class RainbowTrout:
+
+    def __init__(self):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = "Lucky Charm"
+        self.species = "fish"
+        self.swimming = True
+        self.date_added =  date.today()
+        
+class Bluegrill:
+
+    def __init__(self):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = "Bishop Blue"
+        self.species = "fish"
+        self.swimming = True
+        self.date_added =  date.today()
+        
+        
+class Pumpkinseed:
+
+    def __init__(self):
+        # Establish the properties of each animal
+        # with a default value
+        self.name = "Patrick"
+        self.species = "fish"
+        self.swimming = True
+        self.date_added =  date.today()
